+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/craft-ai/craft-ai-client-js/compare/v1.9.0...HEAD) ##
+### Added ###
+- Introducing `craftai.interpreter.getDecisionRulesProperties` a function extracting the property used in the decision rules of a given decision tree.
+- Introducing `craftai.interpreter.decideFromContextsArray` a function that take decisions on an array of several contexts at once.
+
+### Fixed ###
+- `craftai.interpreter.decide` no longer omits to return the standard deviation if it is exactly 0.
+
+### Deprecated ###
+- `craftai.decide` is deprecated in favor of `craftai.interpreter.decide`.
 
 ## [1.9.0](https://github.com/craft-ai/craft-ai-client-js/compare/v1.8.0...v1.9.0) - 2017-08-22 ##
 ### Added ###
@@ -92,7 +101,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added ###
 - Introduce `client.getSharedAgentInspectorUrl` and `client.deleteSharedAgentInspectorUrl` to, respectively, create (or retrieve) and delete the shared agent inspector url.
 ### Deprecated ###
-- `client.getAgentInspectorUrl` is deprecated in favor of `client.getSharedAgentInspectorUrl`.x
+- `client.getAgentInspectorUrl` is deprecated in favor of `client.getSharedAgentInspectorUrl`.
 ### Changed ###
 - Improve the release scripts.